backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
google-generativeai==0.3.2
pytest==8.2.0
httpx>=0.24.0
python-dotenv==1.0.0